Add base layout
