+++ app/src/main/java/com/example/a2022swproject/mainFunction/userBoard/BoardViewModel/SearchResultViewModel.java
@@ -1,0 +1,4 @@
+package com.example.a2022swproject.mainFunction.userBoard.BoardViewModel;
+
+public class SearchResultViewModel {
+}
