+++ app/src/main/java/com/example/a2022swproject/mainFunction/userBoard/BoardViewModel/WritingBoardViewModel.java
@@ -1,4 +1,19 @@
 package com.example.a2022swproject.mainFunction.userBoard.BoardViewModel;
 
-public class WritingBoardViewModel {
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class WritingBoardViewModel extends ViewModel {
+
+    private final MutableLiveData<String> mText;
+
+    public WritingBoardViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is dashboard fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
 }
