+++ app/src/main/res/layout/fragment_homemap.xml
@@ -6,10 +6,20 @@
     android:layout_height="match_parent"
     tools:context=".mainFunction.homeMap.HomeMapFragment">
 
+
     <!--네이버 지도-->
     <com.naver.maps.map.MapView
-        android:id = "@+id/map_view"
+        android:id="@+id/map_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
+    <SearchView
+        android:layout_width="346dp"
+        android:layout_height="44dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.492"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.081" />
 </androidx.constraintlayout.widget.ConstraintLayout>
