+++ app/src/main/res/layout/fragment_userinformation.xml
@@ -7,16 +7,79 @@
     tools:context=".mainFunction.userInformation.UserInformationFragment">
 
     <TextView
-        android:id="@+id/text_notifications"
-        android:layout_width="match_parent"
+        android:id="@+id/userInformation_tv_userName"
+        android:layout_width="179dp"
+        android:layout_height="25dp"
+        android:layout_marginEnd="56dp"
+        android:text=" 로그인을 해주세요"
+        android:textSize="20sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/userInformation_cv_cardviewMember" />
+
+    <TextView
+        android:id="@+id/userInformation_tv_receiveNumber"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:text="0"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="@+id/userInformation_tv_boardNumber"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toEndOf="@+id/userInformation_tmptv_receiveNumber"
+        app:layout_constraintTop_toBottomOf="@+id/userInformation_tv_boardNumber" />
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/userInformation_cv_cardviewMember"
+        android:layout_width="94dp"
+        android:layout_height="86dp"
+        android:elevation="6dp"
+        app:cardCornerRadius="100dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toStartOf="@+id/userInformation_tv_userName"
+        app:layout_constraintHorizontal_bias="0.451"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.046">
+
+        <ImageView
+            android:id="@+id/userInformation_iv_userPicture"
+            android:layout_width="match_parent"
+            android:layout_height="98dp"
+            tools:src="@tools:sample/avatars" />
+    </androidx.cardview.widget.CardView>
+
+    <TextView
+        android:id="@+id/userInformation_tmptv_receiveNumber"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:text="받아온 가구 수:"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="@+id/userInformation_tmptv_boardNumber"
+        app:layout_constraintTop_toBottomOf="@+id/userInformation_tmptv_boardNumber" />
+
+    <TextView
+        android:id="@+id/userInformation_tv_boardNumber"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="12dp"
+        android:text="0"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="@+id/userInformation_tv_userName"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toEndOf="@+id/userInformation_tmptv_boardNumber"
+        app:layout_constraintTop_toBottomOf="@+id/userInformation_tv_userName" />
+
+    <TextView
+        android:id="@+id/userInformation_tmptv_boardNumber"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="12dp"
+        android:text="게시글 작성 건수:"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="@+id/userInformation_tv_userName"
+        app:layout_constraintTop_toBottomOf="@+id/userInformation_tv_userName" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
