+++ app/src/main/res/layout/fragment_writingboard.xml
@@ -1,6 +1,131 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".mainFunction.userBoard.BoardFragment.WritingBoardFragment">
 
+
+    <EditText
+        android:id="@+id/editTextTextPersonName2"
+        android:layout_width="355dp"
+        android:layout_height="45dp"
+        android:ems="10"
+        android:inputType="textPersonName"
+        android:textSize="16sp"
+        android:hint="제목을 입력해주세요!"
+        app:layout_constraintStart_toStartOf="@+id/writingBoard_tv_title"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_title" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_upload"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="게시글 업로드"
+        android:textColor="#000000"
+        android:textSize="24sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.058"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.01" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_uploadDescrpition"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="14dp"
+        android:text="쓸만한 가구를 업로드 해주세요."
+        app:layout_constraintStart_toStartOf="@+id/writingBoard_tv_upload"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_upload" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:text="제목"
+        android:textColor="#292626"
+        android:textSize="18sp"
+        app:layout_constraintStart_toStartOf="@+id/writingBoard_tv_uploadDescrpition"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_uploadDescrpition" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_location"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:text="위치"
+        android:textColor="#292626"
+        android:textSize="18sp"
+        android:textStyle="bold"
+        app:layout_constraintStart_toStartOf="@+id/editTextTextPersonName2"
+        app:layout_constraintTop_toBottomOf="@+id/editTextTextPersonName2" />
+
+    <EditText
+        android:id="@+id/editTextTextPersonName"
+        android:layout_width="316dp"
+        android:layout_height="45dp"
+        android:ems="10"
+        android:inputType="textPersonName"
+        android:hint = "버튼을 통해 현재 위치를 받아올 수 있습니다."
+        android:textSize="16sp"
+        app:layout_constraintStart_toStartOf="@+id/writingBoard_tv_location"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_location" />
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="355dp"
+        android:layout_height="171dp"
+        app:layout_constraintStart_toStartOf="@+id/writingBoard_tv_ex1"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_ex1"
+        app:srcCompat="@android:drawable/ic_menu_report_image" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_ex2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="사진 입력은 게시글 필수 항목입니다."
+        android:textColor="#25D495"
+        app:layout_constraintStart_toStartOf="@+id/imageView"
+        app:layout_constraintTop_toBottomOf="@+id/imageView" />
+
+    <Button
+        android:id="@+id/button"
+        android:layout_width="355dp"
+        android:layout_height="81dp"
+        android:text="게시글 작성"
+        android:textSize="34sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_ex2"
+        app:layout_constraintVertical_bias="0.202" />
+
+    <TextView
+        android:id="@+id/writingBoard_tv_ex1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="24dp"
+        android:text="사진을 업로드 해주세요"
+        app:layout_constraintStart_toStartOf="@+id/editTextTextPersonName"
+        app:layout_constraintTop_toBottomOf="@+id/editTextTextPersonName" />
+
+    <Button
+        android:id="@+id/button3"
+        android:layout_width="37dp"
+        android:layout_height="45dp"
+        android:isScrollContainer="false"
+        app:icon="@android:drawable/ic_menu_mylocation"
+        app:iconGravity="textTop"
+        app:iconPadding="4dp"
+        app:iconTint="#FFFFFF"
+        app:layout_constraintEnd_toEndOf="@+id/editTextTextPersonName2"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toEndOf="@+id/editTextTextPersonName"
+        app:layout_constraintTop_toBottomOf="@+id/writingBoard_tv_location" />
 </androidx.constraintlayout.widget.ConstraintLayout>
